+++ src/main/scala/com/github/plume/oss/domain/package.scala
@@ -28,14 +28,6 @@ package object domain {
     .addModule(DefaultScalaModule)
     .build()
 
-  case class DataFlowCacheConfig(
-      dataFlowCacheFile: Option[Path] = Some(Paths.get("dataFlowCache.cbor")),
-      compressDataFlowCache: Boolean = true,
-      maxCallDepth: Int = 2,
-      maxCachedPaths: Int = 1_000,
-      shareCacheBetweenTasks: Boolean = false
-  )
-
   /** Given an object and a path, will serialize the object to the given path.
     * @param o object to serialize.
     * @param p path to write serialized data to.