+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.3] - 2022-04-05
+
+### Changed
+
+- Added ability to disable cache sharing on `flowsBetween`.
+- Simplified `EngineContext` and `Semantics` on `OverflowDbDriver` startup.
+- `OverflowDbDriver::methodSemantics` now public.
+
+## [1.2.2] - 2022-04-05
+
+### Changed
+
+- Upgraded Joern version to include configurations that disable cache sharing.
+
 ## [1.2.1] - 2022-04-01
 
 ### Changed